changing to new use case
