+++ lib/mergometer.rb
@@ -1,6 +1,6 @@
 require "dotenv/load"
 require "mergometer/configuration"
-require "mergometer/reports_generator"
+require "mergometer/stream"
 require "mergometer/version"
 
 module Mergometer
